+++ packages/jm-ms-http-client/lib/browser/index.js
@@ -16,6 +16,4 @@ const adapter = {
   }
 }
 
-let $ = mdl(adapter)
-$.createModule = mdl // deprecated
-module.exports = $
+module.exports = mdl(adapter)