+++ packages/jm-ms-ws-client/lib/browser/index.js
@@ -1,5 +1,4 @@
 const Adapter = require('./ws')
 const mdl = require('../mdl')
 
-let $ = mdl(Adapter)
-module.exports = $
+module.exports = mdl(Adapter)