+++ packages/jm-ms-http-client/History.md
@@ -1,4 +1,8 @@
 # History
 
+## 2.2.0 / 2020-02-27
+- upgrade dependencies
+- export function without async
+
 ## 2.0.0 / 2018-07-01
 - create